+++ src-tauri/src/models/user_project.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct UserProject {
+  pub name: String,
+  pub path: String,
+}